cr: add report class
